+++ agent/azure-monitor-exporter/src/main/java/com/azure/monitor/opentelemetry/exporter/implementation/LogDataMapper.java
@@ -203,8 +203,7 @@ public class LogDataMapper {
       AbstractTelemetryBuilder telemetryBuilder, Attributes attributes) {
     String invocationId = attributes.get(AiSemanticAttributes.AZ_FN_INVOCATION_ID);
     if (invocationId != null) {
-      telemetryBuilder.addProperty(
-          "InvocationId", attributes.get(AiSemanticAttributes.AZ_FN_INVOCATION_ID));
+      telemetryBuilder.addProperty("InvocationId", invocationId);
     }
     String processId = attributes.get(AiSemanticAttributes.AZ_FN_PROCESS_ID);
     if (processId != null) {